NAV: Clean up LiDAR code
