+++ ieee2017_navigation/nodes/mag_calibrator.py
