+++ ieee2017_navigation/package.xml
@@ -4,7 +4,7 @@
   <version>0.0.0</version>
   <description>The ieee2017_navigation package</description>
 
-  <maintainer email="matthew.langford95@gmail.com">Matthew Langford</maintainer>
+  <maintainer email="anthony@iris-systems.net">Anthony Olive</maintainer>
 
   <license>MIT</license>
